functest_requirements.txt
git+https://github.com/PulpQE/pulp-smash.git#egg=pulp-smash
pulpcore-client
pulp-file-client
pulp-certguard-client
pytest
